refactor(llm): split large runtime helpers
