+++ src/main/java/me/golemcore/bot/adapter/outbound/llm/Langchain4jToolArgumentJson.java
@@ -1,0 +1,41 @@
+package me.golemcore.bot.adapter.outbound.llm;
+
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import java.util.Collections;
+import java.util.Map;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+final class Langchain4jToolArgumentJson {
+
+    private static final TypeReference<Map<String, Object>> MAP_TYPE_REF = new TypeReference<>() {
+    };
+
+    private Langchain4jToolArgumentJson() {
+    }
+
+    static String toJson(Map<String, Object> args, ObjectMapper objectMapper) {
+        if (args == null || args.isEmpty()) {
+            return "{}";
+        }
+        try {
+            return objectMapper.writeValueAsString(args);
+        } catch (Exception e) {
+            log.warn("Failed to serialize tool arguments: {}", e.getMessage());
+            return "{}";
+        }
+    }
+
+    static Map<String, Object> parse(String json, ObjectMapper objectMapper) {
+        if (json == null || json.isBlank()) {
+            return Collections.emptyMap();
+        }
+        try {
+            return objectMapper.readValue(json, MAP_TYPE_REF);
+        } catch (Exception e) {
+            log.warn("Failed to parse tool arguments: {}", e.getMessage());
+            return Collections.emptyMap();
+        }
+    }
+}
